+++ cpp/src/scheduler/Utils.h
@@ -21,7 +21,7 @@
 
 namespace zilliz {
 namespace milvus {
-namespace engine {
+namespace scheduler {
 
 uint64_t
 get_current_timestamp();