+++ cpp/src/scheduler/resource/CpuResource.cpp
@@ -21,7 +21,7 @@
 
 namespace zilliz {
 namespace milvus {
-namespace engine {
+namespace scheduler {
 
 std::ostream &operator<<(std::ostream &out, const CpuResource &resource) {
     out << resource.Dump();